Refactored the code by moving the ToDoList component into its own file
