+++ todo-list/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\index.js":"1","C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\App.js":"2","C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\reportWebVitals.js":"3"},{"size":500,"mtime":1609988535492,"results":"4","hashOfConfig":"5"},{"size":1404,"mtime":1609989761874,"results":"6","hashOfConfig":"5"},{"size":362,"mtime":499162500000,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1ajxg8d",{"filePath":"11","messages":"12","errorCount":2,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\index.js",[],["15","16"],"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\App.js",["17","18","19","20"],"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\reportWebVitals.js",[],{"ruleId":"21","replacedBy":"22"},{"ruleId":"23","replacedBy":"24"},{"ruleId":"25","severity":1,"message":"26","line":5,"column":3,"nodeType":"27","messageId":"28","endLine":5,"endColumn":11},{"ruleId":"25","severity":1,"message":"29","line":6,"column":3,"nodeType":"27","messageId":"28","endLine":6,"endColumn":17},{"ruleId":"30","severity":2,"message":"31","line":31,"column":8,"nodeType":"32","endLine":31,"endColumn":23},{"ruleId":"30","severity":2,"message":"33","line":34,"column":10,"nodeType":"32","endLine":34,"endColumn":18},"no-native-reassign",["34"],"no-negated-in-lhs",["35"],"no-unused-vars","'selector' is defined but never used.","Identifier","unusedVar","'useRecoilState' is defined but never used.","react/jsx-no-undef","'TodoItemCreator' is not defined.","JSXIdentifier","'TodoItem' is not defined.","no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\index.js":"1","C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\App.js":"2","C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\reportWebVitals.js":"3","C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\ToDoList.js":"4"},{"size":500,"mtime":1609988535492,"results":"5","hashOfConfig":"6"},{"size":344,"mtime":1609989998025,"results":"7","hashOfConfig":"6"},{"size":362,"mtime":499162500000,"results":"8","hashOfConfig":"6"},{"size":1338,"mtime":1609990000732,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},"1ajxg8d",{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"12"},{"filePath":"17","messages":"18","errorCount":2,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\index.js",[],["19","20"],"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\App.js",[],"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\reportWebVitals.js",[],"C:\\Users\\fatim\\Desktop\\LambdaProjects\\Personal Projects\\recoil-todo-list\\todo-list\\src\\ToDoList.js",["21","22","23","24"],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},{"ruleId":"29","severity":1,"message":"30","line":5,"column":3,"nodeType":"31","messageId":"32","endLine":5,"endColumn":11},{"ruleId":"29","severity":1,"message":"33","line":6,"column":3,"nodeType":"31","messageId":"32","endLine":6,"endColumn":17},{"ruleId":"34","severity":2,"message":"35","line":30,"column":10,"nodeType":"36","endLine":30,"endColumn":25},{"ruleId":"34","severity":2,"message":"37","line":33,"column":12,"nodeType":"36","endLine":33,"endColumn":20},"no-native-reassign",["38"],"no-negated-in-lhs",["39"],"no-unused-vars","'selector' is defined but never used.","Identifier","unusedVar","'useRecoilState' is defined but never used.","react/jsx-no-undef","'TodoItemCreator' is not defined.","JSXIdentifier","'TodoItem' is not defined.","no-global-assign","no-unsafe-negation"]
